Add to prerequisites (links stopped working!)
After this commit internal links e.g. between intro.md and
prerequisites.md stopped working and `bundle exec jekyll serve` failed
with errors about non-existing paths.
So this is a commit just as much intended for me to be able to switch to
a previous commit when links where working and try to debugg what
happened.