Switch to VimTeX-based PDF reader setup
The idea is: use VimTeX's built-in solutions if they work, and only if
they do not should I start messing around with custom solutions.
This makes the article simpler and keeps readers from having to
configure RPC and server names if they do not want to.
As a result, I will slowly trim custom material from this article.