Move ftplugin content to dedicated article
It makes more sense decouple the ftplugin content, which really anyone
dealing with filetype plugins should now, from the more dense Vimscript
content related to functions and key mappings. That way the Vimscript
article can be more of an optional thing for those interested, and
everyone reads the ftplugin article, which might come earlier in the
series, e.g. as part of the prerequisites.