Enclose suggestions in <details> elements (better readability)
Motivation: most users will meet most of the prerequisites, so it makes
sense for the suggestions/crash courses to be hidden by default.
This makes it easier to skim through the article, and only selectively
read those prerequisites you need to.

Implementation: I used the built-in HTML <details> and <summary>
elements, which is a really nice native solution.
The consequence: the entire document is now basically HTML.